rename: Chip 폴더 이름 통일
