+++ components/chips/ColorPalette.tsx
